# v0/openapi/aihub_openapi.yaml
openapi: 3.1.0
info:
  title: ieops-aihub
  version: 0.4.0
  description: |
    polyforge-v3 backend. v3-design.md §5 (schema) + §10.1 (MCP tools)。
    所有 mutating endpoint 强制带 AttemptCredential 三联 (attempt_id, claim_epoch, session_secret)。
servers:
  - url: http://10.146.0.16
    description: GMI 内网默认部署 (design §6.1 推荐 HTTPS for 公网, warn-only on http://非localhost)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: aihub_sk_*
  responses:
    Unauthorized:
      description: 401 — invalid api_key / session_secret / lease
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: 403 — actor mismatch / not authorized for project
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: "404 — resource not found"
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    BadRequest:
      description: 400 — payload schema 校验失败
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Conflict:
      description: 409 — epoch mismatch / lease expired / lock hard_block / busy_not_takeover_eligible
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
    AttemptCredential:
      type: object
      required: [attempt_id, claim_epoch, session_secret]
      properties:
        attempt_id: {type: string, pattern: "^ra_[a-z0-9]+$"}
        claim_epoch: {type: integer, minimum: 1}
        session_secret:
          type: string
          pattern: "^[0-9a-f]{64}$"
          minLength: 64
          maxLength: 64
          description: "raw 32-byte hex (64 chars); server sha256 比对存储的 session_secret_hash (§6.1)"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT_EPOCH_MISMATCH
            - CONFLICT_LEASE_EXPIRED
            - CONFLICT_LEASE_TAKEN_OVER
            - CONFLICT_HARD_BLOCK
            - CONFLICT_BUSY_NOT_TAKEOVER_ELIGIBLE
            - CONFLICT_DUPLICATE_REQUEST
            - BAD_REQUEST
            - NOT_FOUND
            - PAYLOAD_TOO_LARGE
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - UNKNOWN_REMOTE_ERROR
          description: "13 known ErrorCode (§6.1) + UNKNOWN_REMOTE_ERROR forward-compat placeholder; SoT polyforge_v3.aihub.errors.ErrorCode"
        message: {type: string}
        details: {type: object, additionalProperties: true}
    DeclaredResource:
      type: object
      description: |
        Intent-time 形状 (CreateWorkItemRequest / UpdateWorkItemRequest 入)。
        Runtime fields (state/version/last_sha/last_pr_number/error) 不在此 schema 暴露,
        通过 WorkItemDetailResponse.per_repo_state 返 PerRepoState。详见 §5.1 / §7.6。
      required: [type, uri, intent]
      properties:
        type: {type: string, enum: [repo, path, document, section, service, external_ref]}
        uri: {type: string}
        intent: {type: string, enum: [read, write, refactor, delete]}
        base_branch: {type: string, description: "optional; type=repo 才有"}
        task_branch: {type: string, description: "optional; type=repo 才有"}
        worktree_path_hint: {type: string, description: "optional; display 用, 不全局唯一"}
        metadata: {type: object}
    PerRepoState:
      type: object
      description: |
        Runtime per-repo state (§7.6). Server 维护, WorkItemDetailResponse.per_repo_state
        返。客户端不直接 mutate (走 reconcile / scenario.on_wrap)。
      required: [type, uri, state, version]
      properties:
        type: {type: string, enum: [repo, path, document, section, service, external_ref]}
        uri: {type: string}
        intent: {type: string, enum: [read, write, refactor, delete]}
        state:
          type: string
          enum: [declared, prepared, committed, pushed, pr_opened,
                 blocked_merge_conflict, blocked_dirty_worktree, failed, skipped]
          description: "§7.6 per-repo state machine"
        version: {type: integer, minimum: 0, description: "per-repo state CAS 版本"}
        base_branch: {type: string}
        task_branch: {type: string}
        last_sha: {type: string, pattern: "^[0-9a-f]{7,40}$"}
        last_pr_number: {type: integer}
        error: {type: [string, "null"]}
        metadata: {type: object}
    WorkItem:
      type: object
      required: [id, project, scenario, goal, status, priority, labels,
                 reporter_user_id, declared_resources, resources_version,
                 metadata, created_at, updated_at]
      properties:
        id: {type: string, pattern: "^wi_[a-z0-9]+$"}
        project: {type: string}
        scenario: {type: string}
        goal: {type: string}
        status: {type: string, enum: [queued, running, blocked, paused, wrapped, failed]}
        priority: {type: string, enum: [low, normal, high, urgent]}
        labels: {type: array, items: {type: string}}
        reporter_user_id: {type: string}
        current_attempt_id: {type: [string, "null"]}
        declared_resources: {type: array, items: {$ref: '#/components/schemas/DeclaredResource'}}
        resources_version: {type: integer}
        external_share_type: {type: [string, "null"], enum: [jira, github, null]}
        external_share_key: {type: [string, "null"]}
        parent_work_item_id: {type: [string, "null"]}
        metadata: {type: object}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        closed_at: {type: [string, "null"], format: date-time}
    RunAttempt:
      type: object
      required: [id, work_item_id, status, claim_epoch, idempotency_key,
                 lease_until, actor_user_id, api_key_id, actor_display,
                 machine_id, started_at]
      properties:
        id: {type: string, pattern: "^ra_[a-z0-9]+$"}
        work_item_id: {type: string}
        status: {type: string, enum: [running, superseded, wrapped, failed, expired]}
        claim_epoch: {type: integer, minimum: 1}
        idempotency_key: {type: string}
        lease_until: {type: string, format: date-time}
        actor_user_id: {type: string}
        api_key_id: {type: string}
        actor_display: {type: string}
        machine_id: {type: string}
        session_secret_hash: {type: string}
        parent_attempt_id: {type: [string, "null"]}
        prepared_workspace: {type: [object, "null"]}
        started_at: {type: string, format: date-time}
        ended_at: {type: [string, "null"], format: date-time}
...
    AgentEvent:
      type: object
      required: [id, work_item_id, actor_user_id, event_type, payload, created_at]
      properties:
        id: {type: string}
        work_item_id: {type: string}
        run_attempt_id: {type: [string, "null"]}
        actor_user_id: {type: string}
        api_key_id: {type: [string, "null"]}
        event_type: {type: string}
        payload: {type: object}
        pinned: {type: boolean}
        created_at: {type: string, format: date-time}
...
    CreateWorkItemRequest:
      type: object
      required: [project, scenario, goal]
      properties:
        project: {type: string}
        scenario: {type: string}
        goal: {type: string}
        declared_resources: {type: array, items: {$ref: '#/components/schemas/DeclaredResource'}}
        labels: {type: array, items: {type: string}}
        priority: {type: string, enum: [low, normal, high, urgent]}
        external_share_type: {type: string, enum: [jira, github]}
        external_share_key: {type: string}
        parent_work_item_id: {type: string}
        source:
          type: string
          enum: [human, "auto:review", "auto:execute", "auto:debug",
                 "auto:retro", "sync:jira", "sync:github", "admin"]
          description: "SoT 在 polyforge_v3.core.events.SOURCE_VALUES"
        metadata: {type: object}
    UpdateWorkItemRequest:
      allOf:
        - $ref: '#/components/schemas/AttemptCredential'
        - type: object
          required: [patch_payload]
          properties:
            patch_payload: {type: object}
...
    WorkItemDetailResponse:
      type: object
      required: [work_item, current_attempt, recent_events, per_repo_state]
      properties:
        work_item: {$ref: '#/components/schemas/WorkItem'}
        current_attempt: {$ref: '#/components/schemas/RunAttempt'}
        recent_events: {type: array, items: {$ref: '#/components/schemas/AgentEvent'}}
        per_repo_state: {type: array, items: {$ref: '#/components/schemas/PerRepoState'}}
...
    WorkItemListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/WorkItem'}}
        next_cursor: {type: [string, "null"]}
...
paths:
...
  /v1/work_items:
    post:
      operationId: create_work_item
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateWorkItemRequest'}}}
      responses:
        '201':
          description: created
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkItem'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    get:
      operationId: list_work_items
      parameters:
        - {name: project, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
        - {name: label, in: query, schema: {type: string}}
        - {name: user_id, in: query, schema: {type: string}}
        - {name: source, in: query, schema: {type: string}}
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, default: 50, maximum: 200}}
      responses:
        '200':
          description: list
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkItemListResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /v1/work_items/{work_item_id}:
    parameters:
      - {name: work_item_id, in: path, required: true, schema: {type: string}}
    get:
      operationId: get_work_item
      responses:
        '200':
          description: 详细 + recent events + per_repo state
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkItemDetailResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: update_work_item
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateWorkItemRequest'}}}
      responses:
        '200':
          description: updated
          content: {application/json: {schema: {$ref: '#/components/schemas/WorkItem'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...